// front/src/WeatherModules/VisualCrossing/types.ts
// Generated by https://quicktype.io

export interface WeatherData {
	year: number
	queryCost: number
	latitude: number
	longitude: number
	resolvedAddress: string
	address: string
	timezone: string
	tzoffset: number
	days: Day[]
}

export interface Day {
	datetime: string
	tempmax: number
	tempmin: number
	temp: number
}